requirements.txt
aiofiles==0.8.0
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.2
asyncio==3.4.3
Babel==2.10.3
bcrypt==4.0.0
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iso8601==1.0.2
Jinja2==3.1.2
Mako==1.2.1
mariadb==1.1.4
MarkupSafe==2.1.1
packaging==21.3
pendulum==2.1.2
pydantic==1.9.2
pyparsing==3.0.9
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.2.1
pytzdata==2020.1
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
wrapt==1.14.1
